Wire PacketProcessor send/receive hooks and fragment size
- Send side: write_outgoing called before ENet header, packet_size includes proc header
- Receive side: skip processor bytes, validate_incoming after peer lookup
- Fragment size: subtract processor header_size from mtu in enet_peer_send
- Fix validate_incoming signature to use u16 matching C struct field

+++ src/packet_processor.rs
@@ -34,5 +34,5 @@ pub trait PacketProcessor: Send {
     /// `port` is the host socket's local port.
     /// `peer_reserved` is the current value of the sending peer's reserved
     /// field (0 initially, then the value returned by the previous call).
-    fn validate_incoming(&mut self, header: &[u8], port: u16, peer_reserved: u32) -> Option<u32>;
+    fn validate_incoming(&mut self, header: &[u8], port: u16, peer_reserved: u16) -> Option<u16>;
 }
